@phenomic/plugin-renderer-react: magically consume Html from a file
**Breaking change**: ``createApp()`` don't take a second argument anymore. Html component needs to be placed in a ``Html.js`` file near ``App.js``.
