+++ examples/react-app-with-styled-components/Html.js
@@ -1,0 +1,16 @@
+import React from "react";
+import { ServerStyleSheet } from "styled-components";
+
+export default ({ App, render }) => {
+  const sheet = new ServerStyleSheet();
+  const { Body } = render(sheet.collectStyles(<App />));
+
+  return (
+    <html>
+      <head>
+        {sheet.getStyleElement()}
+      </head>
+      <Body />
+    </html>
+  );
+};
