+++ packages/plugin-renderer-react/src/createApp.js
@@ -1,19 +1,14 @@
 import render from "./render/client";
 
 export type AppType = {
-  routes: React$Element<any>,
-  Html?: (props: PhenomicHtmlPropsType) => React$Element<*>
+  routes: React$Element<any>
 };
 
-export default (
-  routes: () => React$Element<any>,
-  Html?: PhenomicHtmlType
-): AppType => {
+export default (routes: () => React$Element<any>): AppType => {
   if (typeof window !== "undefined") {
     render(routes);
   }
   return {
-    routes: routes(),
-    Html
+    routes: routes()
   };
 };
