+++ packages/preset-react-app/docs/getting-started/6.md
@@ -45,10 +45,16 @@ const BlogPost = ({ isLoading, page }) =>
       page.node &&
       <article>
         <Head>
-          <title>{page.node.title}</title>
+          <title>
+            {page.node.title}
+          </title>
         </Head>
-        <h1>{page.node.title}</h1>
-        <BodyRenderer>{page.node.body}</BodyRenderer>
+        <h1>
+          {page.node.title}
+        </h1>
+        <BodyRenderer>
+          {page.node.body}
+        </BodyRenderer>
       </article>}
     <footer>
       {/* ... */}
@@ -56,8 +62,8 @@ const BlogPost = ({ isLoading, page }) =>
   </div>;
 ```
 
-In order to have the injection of meta into your static html files, you will need
-a custom html template.
+In order to have the injection of meta into your static html files,
+you will need a custom html template.
 
 ⚠️ _Default ``Html`` component is very rudimentary and should be adapted
 if you care about SEO_.
@@ -66,7 +72,8 @@ Here is what you need to add if you want to includes all
 ``react-helmet`` output:
 
 ```js
-const Html = ({ App, render }) => {
+// Html.js
+export default ({ App, render }) => {
   const { Body } = render(<App />);
   const helmet = Head.renderStatic();
   return (
@@ -86,19 +93,9 @@ const Html = ({ App, render }) => {
 };
 ```
 
-You also need to specify your custom ``Html`` component to ``createApp()`` call:
+**This file ``Html.js`` file should be placed next to ``App.js``**
 
-```diff
-export default createApp(() =>
-  <Router history={browserHistory}>
-    { /* ... */ }
--  </Router>
-+  </Router>,
-+  Html
-)
-```
-
-Now your static render should include all meta generated by ``react-helmet``
+Now your static render should include all meta generated by ``react-helmet``.
 
 **To know more about all the things you can do with ``react-helmet``,
 please read [the documentation](https://github.com/nfl/react-helmet#readme).**
