+++ packages/preset-react-app/docs/getting-started/8.md
@@ -13,12 +13,10 @@ import { createApp } from '@phenomic/preset-react-app/lib/client';
 
 // ...
 
-export default createApp(
-  () =>
-    <Router history={browserHistory}>
-      {/* ... */}
-    </Router>,
-  Html
+export default createApp(() =>
+  <Router history={browserHistory}>
+    {/* ... */}
+  </Router>
 );
 ```
 
@@ -34,7 +32,7 @@ const routes = () =>
     {/* ... */}
   </Router>;
 
-export default createApp(routes, Html);
+export default createApp(routes);
 
 if (module.hot) {
   module.hot.accept(() => renderApp(routes));
